third_party/WebKit/LayoutTests/fast/dom/setCustomValidity-null-undefined.html

<!DOCTYPE HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

function test1(elementType) {
    test(function() {
        var element = document.createElement(elementType);
        element.setCustomValidity('');
        assert_equals(element.validationMessage, '');
        element.setCustomValidity(null);
        assert_equals(element.validationMessage, 'null');
        element.setCustomValidity(undefined);
        assert_equals(element.validationMessage, 'undefined');
    }, elementType + ' element test for setCustomValidity(), null and undefined arguments.');
}

function test2(elementType) {
    test(function() {
        var element = document.createElement(elementType);
        element.setCustomValidity('');
        assert_equals(element.validationMessage, '');
        element.setCustomValidity(null);
        assert_equals(element.validationMessage, '');
        element.setCustomValidity(undefined);
        assert_equals(element.validationMessage, '');
    }, elementType + ' element test for setCustomValidity(), null and undefined arguments.');
}

test1('button');
test1('input');
test1('textarea');
test1('select');

test2('output');
test2('fieldset');
test2('object');
</script>
</html>
